# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Reimbursement Prep API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects/{project_id}/receipts:
    get:
      summary: List receipts
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
        - $ref: ./openapi_components_core.yaml#/components/parameters/matched
        - $ref: ./openapi_components_core.yaml#/components/parameters/ocr_status
        - $ref: ./openapi_components_core.yaml#/components/parameters/cursor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/ReceiptList
    post:
      summary: Create receipt placeholder
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ReceiptCreateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Receipt
  /receipts/{receipt_id}/upload-url:
    post:
      summary: Create upload url
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/UploadUrlRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/UploadUrlResponse
  /receipts/{receipt_id}/complete:
    post:
      summary: Complete upload
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ReceiptCompleteRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Receipt
  /receipts/{receipt_id}:
    patch:
      summary: Update receipt fields
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ReceiptUpdateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Receipt
    delete:
      summary: Delete receipt
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      responses:
        "200":
          description: OK
  /receipts/{receipt_id}/match:
    patch:
      summary: Match or unmatch receipt
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ReceiptMatchRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Receipt
  /receipts/{receipt_id}/download-url:
    post:
      summary: Create receipt download url
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/receipt_id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/DownloadUrlResponse
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT